Check if files referenced in source files exist in a current dir

It goes through the source files (.R/.qmd etc.),

It identifies data files (.csv, .xlsx) read or written

Search on the system if these files exist.

Usage

check_referenced_files(path = ".", quiet = FALSE)

Arguments

path

a directory to search for

quiet

Whether it should print messages?

Value

A logical

Details

Still WIP, so you can add code for false positive as needed.

To find genuine referenced files, we exclude different paths

Those created with fs::path() or file.path() or glue::glue()

Those that are checked for fs::file_exists(), file.exists()

Deleted with fs::file_delete(), unlink()