Opens a RStudio project in a new session

If not specified, will generate hyperlinks that call usethis::proj_activate(). proj_switch() looks at options(reuseme.reposdir).

Usage

proj_switch(proj = NA, new_session = TRUE)

Arguments

proj

the name of a project located in the default locations or NA

new_session

Should open in a new session?

Value

Single logical value indicating if current session is modified.

See also