Read and clean names and set variable labels

Reads data from file, but clean names with janitor::make_clean_names().

Usage

read_clean(file, ...)

Arguments

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector).

Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded and decompressed.

Literal data is most useful for examples and tests. To be recognised as literal data, the input must be either wrapped with I(), be a string containing at least one new line, or be a vector containing at least one string with a new line.

Using a value of clipboard() will read from the system clipboard.

Details

It keeps the variable labels attributes to have improved support in the RStudio IDE, with gt tables and with ggplots (>= 3.6.0).

Examples

read_clean(system.file("extdata",  "challenge.csv", package = "readr"))
#> Rows: 2000 Columns: 2
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> dbl  (1): x
#> date (1): y
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 2,000 × 2
#>        x y     
#>    <dbl> <date>
#>  1   404 NA    
#>  2  4172 NA    
#>  3  3004 NA    
#>  4   787 NA    
#>  5    37 NA    
#>  6  2332 NA    
#>  7  2489 NA    
#>  8  1449 NA    
#>  9  3665 NA    
#> 10  3863 NA    
#> # ℹ 1,990 more rows