Get a widget that displays diff between data frame

Thin wrapper around diffviewer::visual_diff().

Usage

visual_diff_df(old, new, width = NULL, height = NULL)

Arguments

old, new

Pair of data frames to compare

width, height

Output size

Value

A HTML widget

Details

Also writes to tempfiles to make sure it works

Examples

mtcars2 <- mtcars

mtcars2[1, 1] <- NA
visual_diff_df(old = mtcars, new = mtcars2)