Elegant wrapper around filter and pull

It can be very useful when trying to extract a value from somewhere, and you have one col that represents the unique id.

Usage

extract_cell_value(
  data,
  var,
  filter,
  name = NULL,
  length = NULL,
  unique = FALSE
)

Arguments

data

A data.frame

var

A variable specified as:

a literal variable name

a positive integer, giving the position counting from the left

a negative integer, giving the position counting from the right.

The default returns the last column (on the assumption that's the column you've created most recently).

This argument is taken by expression and supports quasiquotation (you can unquote column names and column locations).

filter

the filter

name

The variable for the name (by default, will look for rownames), can be quoted (safer).

length

A fixed length to check for the output

unique

A logical. Should return unique values?

Value

A (named) character vector (if name is specified)

Examples

# extract the skin_color for C-3PO
library(magrittr)
extract_cell_value(
  data = dplyr::starwars,
  var = skin_color,
  filter = name == "C-3PO",
  length = 1 # ensure the length will be 1.
)
#> [1] "gold"
# will return a named vector of mpg (as mtcars has rownames.)
mtcars %>%
  extract_cell_value(
    var = mpg,
    filter = vs == 0
  )
#>           Mazda RX4       Mazda RX4 Wag   Hornet Sportabout          Duster 360 
#>                21.0                21.0                18.7                14.3 
#>          Merc 450SE          Merc 450SL         Merc 450SLC  Cadillac Fleetwood 
#>                16.4                17.3                15.2                10.4 
#> Lincoln Continental   Chrysler Imperial    Dodge Challenger         AMC Javelin 
#>                10.4                14.7                15.5                15.2 
#>          Camaro Z28    Pontiac Firebird       Porsche 914-2      Ford Pantera L 
#>                13.3                19.2                26.0                15.8 
#>        Ferrari Dino       Maserati Bora 
#>                19.7                15.0 

# Extract hair color for all people
extract_cell_value(
  data = dplyr::starwars,
  var = skin_color,
  filter = TRUE,
  name = "name" # ensure it is a named vector that corresponds to their unique ID
)
#>        Luke Skywalker                 C-3PO                 R2-D2 
#>                "fair"                "gold"         "white, blue" 
#>           Darth Vader           Leia Organa             Owen Lars 
#>               "white"               "light"               "light" 
#>    Beru Whitesun lars                 R5-D4     Biggs Darklighter 
#>               "light"          "white, red"               "light" 
#>        Obi-Wan Kenobi      Anakin Skywalker        Wilhuff Tarkin 
#>                "fair"                "fair"                "fair" 
#>             Chewbacca              Han Solo                Greedo 
#>             "unknown"                "fair"               "green" 
#> Jabba Desilijic Tiure        Wedge Antilles      Jek Tono Porkins 
#>    "green-tan, brown"                "fair"                "fair" 
#>                  Yoda             Palpatine             Boba Fett 
#>               "green"                "pale"                "fair" 
#>                 IG-88                 Bossk      Lando Calrissian 
#>               "metal"               "green"                "dark" 
#>                 Lobot                Ackbar            Mon Mothma 
#>               "light"        "brown mottle"                "fair" 
#>          Arvel Crynyd Wicket Systri Warrick             Nien Nunb 
#>                "fair"               "brown"                "grey" 
#>          Qui-Gon Jinn           Nute Gunray         Finis Valorum 
#>                "fair"       "mottled green"                "fair" 
#>         Jar Jar Binks          Roos Tarpals            Rugor Nass 
#>              "orange"                "grey"               "green" 
#>              Ric Olié                 Watto               Sebulba 
#>                "fair"          "blue, grey"           "grey, red" 
#>         Quarsh Panaka        Shmi Skywalker            Darth Maul 
#>                "dark"                "fair"                 "red" 
#>           Bib Fortuna           Ayla Secura              Dud Bolt 
#>                "pale"                "blue"          "blue, grey" 
#>               Gasgano        Ben Quadinaros            Mace Windu 
#>         "white, blue" "grey, green, yellow"                "dark" 
#>          Ki-Adi-Mundi             Kit Fisto             Eeth Koth 
#>                "pale"               "green"               "brown" 
#>            Adi Gallia           Saesee Tiin           Yarael Poof 
#>                "dark"                "pale"               "white" 
#>              Plo Koon            Mas Amedda          Gregar Typho 
#>              "orange"                "blue"                "dark" 
#>                 Cordé           Cliegg Lars     Poggle the Lesser 
#>               "light"                "fair"               "green" 
#>       Luminara Unduli         Barriss Offee                 Dormé 
#>              "yellow"              "yellow"               "light" 
#>                 Dooku   Bail Prestor Organa            Jango Fett 
#>                "fair"                 "tan"                 "tan" 
#>            Zam Wesell       Dexter Jettster               Lama Su 
#> "fair, green, yellow"               "brown"                "grey" 
#>               Taun We            Jocasta Nu         Ratts Tyerell 
#>                "grey"                "fair"          "grey, blue" 
#>                R4-P17            Wat Tambor              San Hill 
#>         "silver, red"         "green, grey"                "grey" 
#>              Shaak Ti              Grievous               Tarfful 
#>    "red, blue, white"        "brown, white"               "brown" 
#>       Raymus Antilles             Sly Moore            Tion Medon 
#>               "light"                "pale"                "grey" 
#>                  Finn                   Rey           Poe Dameron 
#>                "dark"               "light"               "light" 
#>                   BB8        Captain Phasma         Padmé Amidala 
#>                "none"             "unknown"               "light"