Print interactive outline of file sections

RStudio project, or directories This will fail if you are trying to map an unsaved file.

Usage

file_outline(
  regex_outline = NULL,
  path = active_rs_doc(),
  work_only = TRUE,
  alpha = FALSE,
  dir_common = NULL,
  print_todo = TRUE,
  recent_only = FALSE,
  width = cli::console_width(),
  n_colors = NULL
)

proj_outline(
  regex_outline = NULL,
  proj = proj_get2(),
  work_only = TRUE,
  dir_tree = FALSE,
  alpha = FALSE,
  recent_only = FALSE
)

dir_outline(
  regex_outline = NULL,
  path = ".",
  work_only = TRUE,
  dir_tree = FALSE,
  alpha = FALSE,
  recent_only = FALSE
)

Arguments

regex_outline

A string or regex to search for in the outline

path, proj

A character vector of file paths, a project. Defaults to active file, project or directory. rstudioapi::documentPath()

work_only

If TRUE, (the default), will only show you work items first. Set to FALSE if you want to see the full outline. WORK will combine with regex_outline

alpha

Whether to show in alphabetical order

dir_common

(Do not use it)

print_todo

Should include TODOs in the file outline? If FALSE, will print a less verbose output with sections.

recent_only

Show outline for recent files

width

Width (internal)

n_colors

Number colours (Internal)

dir_tree

If TRUE, will print the fs::dir_tree() or non-R files in the directory

Value

A reuseme_outline object that contains the information. Inherits tbl_df.

A symbol will show for recently modified files.

Details

If work_only is set to TRUE, the function will only return outline of the # WORK comment in path. work_only = TRUE will have an effect on regex_outline. These functions are more By default

file_outline() prints the outline the active document if in RStudio

proj_outline() prints the outline of the active project if in RStudio

dir_outline() prints the outline of the active working directory by default or

proj_* and dir_ call file_outline.

The parser is very opinioneted and is not very robust as it is based on regexps. For a better file parser, explore other options, like lightparser, {roxygen2}

Will show TODO items and will offer a link to mark them as complete

Examples

file <- fs::path_package("reuseme", "example-file", "outline-script.R")
file_outline(path = file)
#> 
#> ── `example-file/outline-script.R` 🦐 
#> `i` Load packages 
#> `i` Wrangle + visualize data 
#> `i` A great title
#> `i` TODO improve this Viz!- `Done✔?`

# Remove TODO
file_outline(path = file, print_todo = FALSE, alpha = TRUE)
#> 
#> ── `example-file/outline-script.R` 🦐 
#> `i` A great title
#> `i` Load packages 
#> `i` Wrangle + visualize data 

# interact with data frame
file_outline(path = file) |> dplyr::as_tibble()
#> # A tibble: 4 × 32
#>   file        line_id content file_ext is_md file_with_many_funct…¹ is_test_file
#>   <chr>         <int> <chr>   <chr>    <lgl> <lgl>                  <lgl>       
#> 1 /home/runn…       1 "# Loa… R        FALSE FALSE                  FALSE       
#> 2 /home/runn…       4 "# Wra… R        FALSE FALSE                  FALSE       
#> 3 /home/runn…       9 "    t… R        FALSE FALSE                  FALSE       
#> 4 /home/runn…      11 "# TOD… R        FALSE FALSE                  FALSE       
#> # ℹ abbreviated name: ¹​file_with_many_functions
#> # ℹ 25 more variables: is_cli_info <lgl>, is_doc_title <lgl>,
#> #   is_chunk_cap <lgl>, is_chunk_cap_next <lgl>, is_test_name <lgl>,
#> #   is_section_title <lgl>, is_tab_or_plot_title <lgl>,
#> #   is_a_comment_or_code <lgl>, is_todo_fixme <lgl>,
#> #   is_section_title_source <lgl>, before_and_after_empty <lgl>,
#> #   n_leading_hash <dbl>, is_second_level_heading_or_more <lgl>, …

if (FALSE) { # interactive()
# These all work on the active file / project or directory.

file_outline()
proj_outline()
dir_outline()
# Like proj_switch(), proj_outline() accepts a project
}